integrate reminder into walle & bug fixes
