+++ package.json
@@ -47,7 +47,7 @@
 		"@types/express": "^5.0.0",
 		"@types/humanize-duration": "^3.27.4",
 		"@types/luxon": "^3.4.2",
-		"@types/node": "^22.13.1",
+		"@types/node": "^22.13.4",
 		"pino-pretty": "^13.0.0",
 		"prisma": "^6.3.1",
 		"typescript": "^5.7.3"
